유저 통합 및 Soft Delete 적용

+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/PortfolioJPARepository.java
@@ -20,5 +20,5 @@ public interface PortfolioJPARepository extends JpaRepository<Portfolio, Long> {
     void deleteByPlanner(Planner planner);
 
     @Query("select p from Portfolio p where p.planner.id = :plannerId")
-    Optional<Portfolio> findByPlannerId(@Param("plannerId") int plannerId);
+    Optional<Portfolio> findByPlannerId(@Param("plannerId") Long plannerId);
 }
